Support hashlib for Python 2.5 and above
Python 2.6 issues a "Warning:" when we use the sha module that
has been deprecated since Python 2.5. The "hashlib" module that
should be used instead is only available in Python 2.5 and above,
so check for the existence and use whichever method we have available
(but prefer hashlib over the older sha module if available).