Retry on error
This implements a way to configure that jobs should be retried if
an error occurs whilst retrieving. The number of maximum retries
can be configured by the user on a per-job base.

If an error occurs and the number of tries has not yet been reached
the program will not display any error. Only after the maximum tries
has been exceeded the error will be reported. This can be used to
e.g. mitigate transient network errors if a host sometimes becomes
temporarily unreachable.

The delay between retries is determined by the time interval between
program executions. This means the interval is in the hands of the
user, if a cronjob is executed a retry will happen every time the
cronjob is executed until the maximum amount of tries is reached.