Raise exception in case of empty job output
This addresses #245 and makes sure an exception is raised (resulting in
an error) when the output from a job is empty, since this is a pretty
good indication for a misconfiguration. The `ignore_empty` disables this
check, making it still possible to monitor empty pages, e.g. in case you
want to watch for an empty page getting filled with content.