Add optional option to ignore network errors in UrlJob
This implements a per-entry configuration option named `ignore_network`
that can be specified if (temporary) network errors should be ignored.
When this option is specified, you won't be notified in case of a
network error. The default is still the same, i.e. if this key is not
present, you'll still get an error and will be notified. This was
discussed in #218.