Implement PDF to text filter 'pdf2text'
A filter 'pdf2text' is created to extract text from a PDF file. It is based on the pdftotext (https://github.com/jalan/pdftotext) library which in turn is based on poppler (https://poppler.freedesktop.org/). Its function only works on bytes, so I also modified code in the job library so that if this filter is specified, a bytes object is returned instead of a str.