Implement smarter logic for trimming Markdown (and hence Matrix) messages.
Instead of simply cutting off a trailing portion of a formatted Markdown
message blindly and thus likely making the formatting invalid (by
chopping of the closing code fence block), we try to be more smart about
it.

First we allocate a maximum length of the report. We render the summary
and deduct its length from the maximum length, arriving at a remaining
length available. From there, we split the remaining length between the
details items (if there are any). If a details item does not use all of
its allocated space, the unused space is passed onto the next item.

Finally, if any of the details items exceeds its allocated length, the
item will be repeatedly line-wise trimmed until it fits. If there is not
enough space for the details at all, the details will be omitted (with
a note informing the user of this occurence).

The resulting Markdown will never exceed the maximum length set for it,
but there are less guarantees on what its final length will be. In
practice, messages often end up very close in size to the maximum
length.